+++ sample-apps/react-native/dogfood/ios/AppDelegate.swift
@@ -77,6 +77,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate, UNUserNotificationCenterD
     }
     
     let uuid = UUID().uuidString
+    let videoIncluded = stream["video"] as? String
+    let hasVideo = videoIncluded == "false" ? false : true
     
     StreamVideoReactNative.registerIncomingCall(cid, uuid: uuid)
     
@@ -89,7 +91,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate, UNUserNotificationCenterD
     RNCallKeep.reportNewIncomingCall(uuid,
                                      handle: createdCallerName,
                                      handleType: "generic",
-                                     hasVideo: true,
+                                     hasVideo: hasVideo,
                                      localizedCallerName: createdCallerName,
                                      supportsHolding: false,
                                      supportsDTMF: false,