+++ sample-apps/react-native/dogfood/src/types/env.d.ts
@@ -1,4 +1,9 @@
-type AppEnvironment = 'pronto' | 'pronto-staging' | 'demo' | 'livestream';
+type AppEnvironment =
+  | 'pronto'
+  | 'pronto-staging'
+  | 'demo'
+  | 'livestream'
+  | 'video-moderation';
 
 declare module '@env' {
   export const REACT_NATIVE_DOGFOOD_APP_ENVIRONMENT: AppEnvironment;